requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
redis==4.6.0
Werkzeug==2.3.7
pdf2image
python-docx==0.8.11
PyPDF2==3.0.1
psycopg2-binary
python-pptx==0.6.23
faker
ffmpeg-python
Pillow

tzdata 

gunicorn